Modified the default ranking such that CSRF can have its correct spot in the lesson ranking scheme.
git-svn-id: http://webgoat.googlecode.com/svn/trunk@82 4033779f-a91e-0410-96ef-6bf7bf53c507

+++  webgoat/main/project/JavaSource/org/owasp/webgoat/lessons/RoleBasedAccessControl/RoleBasedAccessControl.java
@@ -121,7 +121,7 @@ public class RoleBasedAccessControl extends LessonAdapter
 
     private final static String JSP_PATH = "/lessons/" + LESSON_NAME + "/";
 
-    private final static Integer DEFAULT_RANKING = new Integer(120);
+    private final static Integer DEFAULT_RANKING = new Integer(125);
 
     private static Connection connection = null;
